requirements.txt
fastapi
pydantic
uvicorn[standard]
pandas
numpy
dask[complete]
tqdm
gcsfs
fsspec
psycopg2-binary
sqlalchemy
google-cloud-storage
google-cloud-bigquery
boto3[crt]
jsonlines
networkx
matplotlib
scipy
igraph
motor
pymongo
ijson
ipython
seaborn
#multimerge